initial SSH passphrases support
